Anchor::Put

+++ src/types.h
@@ -101,8 +101,8 @@ struct Line
     Line *prev;
     int len;
     int width;
-    long linenumber;      /* on buffer */
-    long real_linenumber; /* on file */
+    int linenumber;      /* on buffer */
+    int real_linenumber; /* on file */
     unsigned short usrflags;
     int size;
     int bpos;
